refactor: collocate logout button as it's used in the private layout
