+++ src/app/(private)/_components/log-out-button.tsx
@@ -1,8 +1,8 @@
 'use client';
 
 import { useMutation } from '@tanstack/react-query';
-import { logoutAction } from '~/app/actions';
 import { Button } from '~/components/ui/button';
+import { logoutAction } from '../_actions/logout';
 
 export function LogoutButton() {
   const mutation = useMutation({ mutationFn: logoutAction });
